fix(tests): inline assert_query_limit in root conftest (avoid apps/*/tests collision)
