+++ conftest.py
@@ -1,7 +1,21 @@
-"""Root conftest: re-exports fixtures from tests/conftest.py so they are
-available across packages/ and apps/ test trees."""
+"""Root conftest — exposes shared fixtures across packages/ and apps/ tests.
+
+Defined inline (no `from tests.conftest import ...`) to avoid pytest
+module-discovery conflicts with `apps/*/tests/` dirs that lack `__init__.py`.
+"""
 from __future__ import annotations
 
-from tests.conftest import assert_query_limit
+from contextlib import AbstractContextManager
 
-__all__ = ["assert_query_limit"]
+import pytest
+
+
+@pytest.fixture
+def assert_query_limit():
+    def _check(response, limit: int = 15) -> None:
+        count = int(response.headers.get("X-Query-Count", "0"))
+        assert count <= limit, f"n_plus_1_violation count={count} limit={limit}"
+    return _check
+
+
+__all__ = ["assert_query_limit", "AbstractContextManager"]
